Stop (we hope) vatstore kernel test from flaking in CI (#450)
Adds a loopback message inside the vatstore test test vat to ensure that
state quiesces before the kernel exits. Such a loopback is the general
practice in the various other kernel integration tests but I didn't
think it was needed here but apparently it is -- or so we shall see.
It's a hypothesis...

BTW, also astonished to discover that vitest's
`expect(x).toStrictEqual(y)` is sensitive to Map insertion order. Who
possibly thought that was a good idea?