refactor: Begin using JSON-RPC for internal messages (#451)
Ref: #281 #369 

This is the first step toward simplifying our internal message types. We
will accomplish this simplifications by making our transports dumber, in
that their only expectation is that every message is a valid JSON-RPC
message. All other validation will be relegated to the RPC call and
method implementation sites.

MetaMask itself uses JSON-RPC for all of its internal (and most
external) messages. To this end, it has a mature stack for handling
JSON-RPC messages and requests/responses, centered around
[@metamask/json-rpc-engine](https://github.com/metamask/core/tree/main/packages/json-rpc-engine).
Here, we introduce `JsonRpcEngine` and JSON-RPC for UI <-> kernel
messages only. This will grow to include all internal messages.

`JsonRpcEngine` is somewhat idiosyncratic (read: has some dumb stuff
we'd like to remove), but is ultimately pretty straightforward to work
with. See its documentation for details. One of the main implications is
for our message types. Where they previously looked e.g. like this:

```json5
// Request
{
  "method": "ping",
  "params": null,
}

// Response
{
  "method": "ping",
  "params": "pong", // or error
}
```

They will now look like this:
```json5
// Request
{
  "id": "0xdeadbeef", // string or number, we'll always use strings for simplicity
  "jsonrpc": "2.0", // required by spec
  "method": "ping",
  "params": [], // "params" must be a "structured value", i.e. object or array
}

// Success response
{
  "id": "0xdeadbeef", // or whatever
  "jsonrpc": "2.0", // required by spec
  "result": "pong",
}

// Error response
{
  "id": "0xdeadbeef", // or whatever
  "jsonrpc": "2.0", // required by spec
  "error": { ... },
}
```

Ultimately, callers and method implementers will never have to concern
themselves with anything other than the `method` and `params` for calls,
and `result` / `error` for responses. The biggest differences from our
existing types are that:
- `params` must be either an array or an object—think positional or
named—and can therefore not be `null`.
- responses must have a `result` or `error` property, never `params`.

This PR will be followed immediately by the next pipeline to receive the
JSON-RPC treatment, and culminate in a rewrite of our RPC method
implementations and their types.