Handle vat resumption on vat or kernel restart (#448)
These changes make it possible to restart a vat (via the control panel)
or restart the kernel (e.g., via a browser restart) without losing
persistent vat state and allowing pre-existing vats to continue running.
Also, when a vat is actually terminated (as opposed to simply shut
down), all the relevant persistent kernel state is now expunged from the
database.

Along the way this cleaned up a lot of the incidental logic concerning
the various forms of restart that are driven from the control panel. It
also ended up pulling in a fix to a nasty bug in the way the kernel
handled promise message queues, and which got in the way of some of the
testing. Also fixed the way the database is set up for kernel
integration tests, so that concurrent tests wouldn't interfere with each
other (this also got in the way of testing).

Closes #437