feat: Add `rpc-methods` package (#474)
Progresses: #369

Introduces a new package, `@ocap/rpc-methods`, whose primary affordances
are two: `RpcClient` and `RpcService`. These classes are intended to
represent the call-er and call-ee ends of an RPC relationship. In brief,
you:
1. Call RPC methods using `RpcClient.call()`
2. pipe the resulting payload to `RpcService.execute()`
3. and pipe the result back to the caller.

These two classes consume "method" and "handler" specifications. The
idea is that you specify your methods individually, gather these
specifications together into a "specification" or "handler" record, and
then let the rest of your RPC types follow from there.

Our first target for this pattern is the "kernel control methods", where
we are able to replace the `MessageResolver` and `CommandRegistry`
abstractions, as well as the hated request / response structs. See
`packages/extension/src/kernel-integration/handlers` for where most of
this action occurred. Although this PR consists of equal amounts of
additions and deletions, you'll note the kernel control method
implementation in the extension is significantly simplified (+700 /
-1300).

The application's behavior should be virtually unaffected by this
change. Following this PR, we will target other RPC channels in order to
get rid of their request / response structs as well.