refactor: Replace "message" IPC pattern with RpcClient and RpcService (#487)
Closes: #369 

Replaces all kernel and vat "message" types with the new `RpcClient` /
`RpcService` pattern. For related work and the justification for this,
see:
- #369
- #474 
- #481 

Note that, in the course of implementing this refactor, it was
discovered that we are never returning _any_ responses to syscalls
originating from vat supervisors. This PR maintains a provisional
solution using the new pattern, and #488 tracks the work to actually fix
this.