feat(streams): Add stream splitting (#503)
We removed stream multiplexing in #349 because we no longer had a use
for it. With some of the plans we have for our logger, we once again
find ourselves in a position where we'd like to differentiate traffic
over a single stream. Multiplexing was pretty complex, so we'd prefer
not to go back to it if possible. With these requirements in mind, this
PR introduces "stream splitting" via the function `split()` exported
from `@ocap/streams`.

The idea is simple: pass `split()` a duplex stream and a number of type
predicates, and receive an array of streams in return. Each stream
represents the subset of the parent stream's traffic satisfying the
corresponding predicate. Not very efficient for large numbers of
channels or complex predicates, but our usage should be simple enough.