feat: TS version of the libp2p browser-to-browser proof of concept
The previously demonstrated libp2p browser-to-browser communications
demo (found in https://github.com/FUDCo/brow-2-brow) was written in
simple JavaScript since it was derived directly from the thing that
https://docs.libp2p.io/guides/getting-started/webrtc/ walks you through,
which was itself written in simple JS.

This version has been converted to TypeScript, hammered into compliance
with our local eslint rules and other fussinesses, and integrated with
the rest of our repo tooling so that you can build and run it in our
environment.

While this now lives within our repo, it should be functionally
independent of and operationally decoupled from all our other stuff.