feat(kernel): Add removable bootstrap keep‑alive pin (#502)
Closes #494 

This PR implements a `pin` and an `unpin` method for vat root objects,
that external code (tests, control panel) can call to keep a vat or
arbitrary kernel object alive to ensure that references are counted.

A small reorganize of the kernel store methods was required and I added
some tests that were missing.