Fix weird libp2p package version sheer (#505)
I was experiencing previously unencountered TypeScript whinage in my
editor, which turned out to be due to a libp2p package versioning issue.
Basically, the TS typing for the APIs used by the libp2p tutorials (from
which this code was derived) don't quite match the types that are
actually declared in the latest released packages, and in my original
conversion to TS I ended up inadvertently combining some old and some
new and the mixture made LSP unhappy (even though it built and ran
fine). But fixing it turned out to be one of those
loose-thread-unravelling-the-whole-sweater things. This PR glues it all
back together into a self-consistent ensemble using the latest libp2p
stuff.