refactor: Prepare for publishing packages to npm (#507)
To publish our packages to `npm`, we need to scope them under the
`@metamask/*` organization. Packages that are to remain private (for now
or in perpetuity) can remain scoped under `@ocap/*`. Packages that need
to be published includes everything that we'll want to use in
production. Consequently, we rename and reconfigure the following:
- `@ocap/kernel` -> `@metamask/ocap-kernel`
- `@ocap/store` -> `@metamask/kernel-store`
- `@ocap/utils` -> `@metamask/kernel-utils`
- `@ocap/shims` -> `@metamask/kernel-shims`
- `@ocap/rpc-methods` -> `@metamask/kernel-rpc-methods`
- `@ocap/errors` -> `@metamask/kernel-errors`
- `@ocap/logger` -> `@metamask/logger`
- `@ocap/streams` -> `@metamask/streams`

Since we're only likely to use the CLI for internal purposes for the
time being, we keep it private for now. The extension and Node.js
runtimes are not likely to ever be published as libraries, and can also
remain private.